# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Book Management API
  version: 1.0.0
  description: 書籍と著者を管理するためのAPI

paths:
  /api/authors:
    post:
      operationId: createAuthor
      summary: 著者を作成
      tags:
        - Authors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
      responses:
        '201':
          description: 著者が正常に作成されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/authors/search:
    get:
      operationId: searchAuthors
      summary: 著者を検索
      tags:
        - Authors
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: 著者名（部分一致）
        - name: dateOfBirthFrom
          in: query
          required: false
          schema:
            type: string
            format: date
          description: 生年月日（から）
        - name: dateOfBirthTo
          in: query
          required: false
          schema:
            type: string
            format: date
          description: 生年月日（まで）
        - name: bookTitle
          in: query
          required: false
          schema:
            type: string
          description: 著作の書籍タイトル（部分一致）
        - name: publicationStatus
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PublicationStatus'
          description: 著作の出版状況
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 検索結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAuthors'

  /api/authors/{id}:
    get:
      operationId: getAuthor
      summary: 著者を取得
      tags:
        - Authors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 著者情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: 著者が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: updateAuthor
      summary: 著者を更新
      tags:
        - Authors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
      responses:
        '200':
          description: 著者が正常に更新されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 著者が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/authors/{id}/books:
    get:
      operationId: getBooksByAuthor
      summary: 著者に紐づく書籍一覧を取得
      tags:
        - Authors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 書籍一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '404':
          description: 著者が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/books:
    post:
      operationId: createBook
      summary: 書籍を作成
      tags:
        - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '201':
          description: 書籍が正常に作成されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/books/search:
    get:
      operationId: searchBooks
      summary: 書籍を検索
      tags:
        - Books
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: 書籍タイトル（部分一致）
        - name: authorName
          in: query
          required: false
          schema:
            type: string
          description: 著者名（部分一致）
        - name: priceFrom
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: 価格（から）
        - name: priceTo
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: 価格（まで）
        - name: publicationStatus
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PublicationStatus'
          description: 出版状況
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 検索結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBooks'

  /api/books/{id}:
    get:
      operationId: getBook
      summary: 書籍を取得
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 書籍情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: 書籍が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: updateBook
      summary: 書籍を更新
      tags:
        - Books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '200':
          description: 書籍が正常に更新されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: バリデーションエラー（出版済み→未出版への変更不可など）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 書籍が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: ページ番号（0始まり）
    SizeParam:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 1ページあたりの件数（最大100）

  schemas:
    Author:
      type: object
      required:
        - id
        - name
        - dateOfBirth
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: 著者ID（UUIDv7）
        name:
          type: string
          description: 著者名
          maxLength: 200
        dateOfBirth:
          type: string
          format: date
          description: 生年月日
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時

    AuthorRequest:
      type: object
      required:
        - name
        - dateOfBirth
      properties:
        name:
          type: string
          description: 著者名
          minLength: 1
          maxLength: 200
        dateOfBirth:
          type: string
          format: date
          description: 生年月日（現在日以前であること）

    Book:
      type: object
      required:
        - id
        - title
        - price
        - publicationStatus
        - authors
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: 書籍ID（UUIDv7）
        title:
          type: string
          description: タイトル
          maxLength: 400
        price:
          type: integer
          description: 価格（0以上であること）
          minimum: 0
        publicationStatus:
          $ref: '#/components/schemas/PublicationStatus'
        authors:
          type: array
          description: 著者一覧（最低1人）
          minItems: 1
          items:
            $ref: '#/components/schemas/Author'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時

    BookRequest:
      type: object
      required:
        - title
        - price
        - publicationStatus
        - authorIds
      properties:
        title:
          type: string
          description: タイトル
          minLength: 1
          maxLength: 400
        price:
          type: integer
          description: 価格（0以上であること）
          minimum: 0
        publicationStatus:
          $ref: '#/components/schemas/PublicationStatus'
        authorIds:
          type: array
          description: 著者IDの配列（最低1人）
          minItems: 1
          items:
            type: string
            format: uuid

    PublicationStatus:
      type: string
      enum:
        - Unpublished
        - Published
      description: 出版状況（出版済みステータスのものを未出版には変更できない）

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: エラーメッセージ
        details:
          type: string
          description: 詳細情報

    PageMeta:
      type: object
      required:
        - page
        - pageSize
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          description: 現在のページ番号（0始まり）
        pageSize:
          type: integer
          description: 1ページあたりの件数
        totalElements:
          type: integer
          format: int64
          description: 総件数
        totalPages:
          type: integer
          description: 総ページ数

    PagedAuthors:
      type: object
      required:
        - content
        - meta
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        meta:
          $ref: '#/components/schemas/PageMeta'

    PagedBooks:
      type: object
      required:
        - content
        - meta
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        meta:
          $ref: '#/components/schemas/PageMeta'